header-button-activate: accept action as synonym for property :action
Previously only `:action' was supported but button.el uses `action'
instead.  Now both properties are supported and `:action' is queried
first because `actions' always defaults to function `ignore'.

Note that this also fixes a bug: the only previously valid property
`:action' did not have a default value, so if it was not explicitly
set an attempt to call nil as a function was made.